REFACTOR :: 폴더 구조 변경
- components domain별로 나눔
- chat 이름 자세히 변경
