+++ components/Chat/ChatList/ChatListSection.js
@@ -1,6 +1,6 @@
 import {ScrollView, StyleSheet} from "react-native";
-import ChatList from "../ChatList/ChatList";
-import ChatListData from "../../data/ChatList.json"
+import ChatList from "./ChatList";
+import ChatListData from "../../../data/ChatList.json"
 
 export default function ChatListSection({navigation, rooms}) {
