+++ components/common/SafeAreaView/MyStatusBar.js
